configurator/requirements.txt
# Metro Digitale Configurator - Dipendenze Python

# GUI Framework
PyQt6>=6.5.0
PyQt6-WebEngine>=6.5.0

# Comunicazione seriale ESP32
pyserial>=3.5

# HTTP requests per Iconify API
requests>=2.31.0

# Manipolazione immagini
Pillow>=10.0.0

# Conversione SVG (opzionale, per rendering icone)
cairosvg>=2.7.0

# Parser formule matematiche
pyparsing>=3.1.0

# Testing
pytest>=7.4.0
pytest-qt>=4.2.0
